+++ src/Kiota.Builder/KiotaBuilder.cs
@@ -1290,7 +1290,7 @@ public partial class KiotaBuilder
             };
             var operationUrlTemplate = currentNode.GetUrlTemplate(operationType);
             if (!operationUrlTemplate.Equals(parentClass.Properties.FirstOrDefault(static x => x.Kind is CodePropertyKind.UrlTemplate)?.DefaultValue?.Trim('"'), StringComparison.Ordinal)
-                && currentNode.HasRequiredQueryParametersAcrossOperations())
+                && currentNode.HasRequiredQueryParametersAcrossOperations())// no need to generate extra strings/templates as optional parameters will have no effect on resolved url.
                 generatorMethod.UrlTemplateOverride = operationUrlTemplate;
 
             var mediaTypes = schema switch